requirements.txt
backcall==0.2.0   
biopython==1.79    
cloudpickle==2.0.0   
cramjam==2.4.0   
cycler==0.10.0  
dask>=2021.10.0
decorator==5.1.0   
dna-features-viewer==3.1.0   
et-xmlfile==1.1.0   
fastparquet==0.7.1   
fsspec==2021.9.0
ipython>=7.31.1  
jedi==0.18.0  
joblib==1.0.1   
kiwisolver==1.3.2   
locket==0.2.1   
matplotlib==3.4.3   
matplotlib-inline==0.1.3   
numpy==1.21.2  
openpyxl==3.0.9   
packaging==21.0    
pandas==1.3.3   
parso==0.8.2   
partd==1.2.0   
pathlib==1.0.1   
patsy==0.5.2   
pexpect==4.8.0   
pickleshare==0.7.5   
Pillow>=9.0.0   
pip>=21.1  
prompt-toolkit==3.0.21  
ptyprocess==0.7.0   
pyarrow==5.0.0   
Pygments==2.10.0  
pyparsing==2.4.7   
python-dateutil==2.8.2   
pytz==2021.1  
PyYAML==5.4.1   
scikit-allel==1.3.5   
scikit-learn==1.0     
scipy==1.7.1   
seaborn==0.11.2  
setuptools==44.0.0  
six==1.16.0  
sklearn==0.0     
statsmodels==0.12.2  
threadpoolctl==2.2.0   
thrift==0.15.0  
toolz==0.11.1  
traitlets==5.1.1   
wcwidth==0.2.5  
